How to ignore a file
Once the file's in a repo, you can't expect the .gitignore to prevent
you from uploading changes.  Apparently the way to "ignore" that is to
tell git's index to assume that file hasn't changed.

I _think_ that'll have to be done by everyone though....  Unless _their_
index assumes it hasn't changed, it'll still be a file to commit.  Hence
the instruction text I'm commiting here ;)