2button demo with options
A 2 button test with selectable options.  This example uses stacked
background images (http://www.css3.info/preview/multiple-backgrounds/).
That doesn't work in IE8.  I need to test if it works on the iPad and
iPod touch (and android).  It works in Chrome 1.3+, Safari 1.0+, Opera
10.5+, IE9+, FF 3.6+.

Demo should be accessible from
http://voteserver-dev.hsc.myproit.com/static/scratch/2Button1/2button1-demo.html
(and the repo, of course).