Clerk Interface - Copy Executable
Created an executable that watches for a window with
"copy-me-to-clerk-soft" in the title, hits "ctrl+c", closes the window,
activates the window with "Granicus" in the title, and hits "ctrl+v".

For now it uses the same html from clerkajaxmotionpull but using a url
flag to add the javascript necessary for the proper element selection.