Clerk Interface - Long Polling Cleanup / misc fix
Changed the Clerk's long poller to only update the votes within each
motion as opposed to refreshing the whole page of motions.

Misc fixes:
- if a motion had no votes, ClerkAjaxMotionPull broke and wouldn't
render that motion at all.  Added a separate variable storing the motion
itself (and associated utility) to grab the motion for display.
- changed long polling timing from .4 seconds to .3 seconds.  No real
change in CPU (apache process used ~0.7% cpu as seen in top in both
scenarios)
- removed unused ClerkAjax view
- removed unused arrays from clerk.html
- removed my 'poo' from the debug_text div in clerk.html
- commented clerkajaxmotionpull.html

Potential to-do:
- added 'array_motions' array to clerk.html in hopes of sending the
motions the page knows about to then be notified of which motions need
to be updated